WIP: NewStateRoot and State API impls
