+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/blocks/impl/BlockImplUtils.java
@@ -1,54 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.hedera.node.app.blocks.impl;
 
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_ACCOUNTS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_ALIASES;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_BLOCK_INFO;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_BLOCK_STREAM_INFO;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_CONGESTION_STARTS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_CONTRACT_BYTECODE;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_CONTRACT_STORAGE;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_ENTITY_ID;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_FILES;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_FREEZE_TIME;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_MIDNIGHT_RATES;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_NETWORK_REWARDS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_NFTS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_NODES;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_PENDING_AIRDROPS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_PLATFORM_STATE;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_ROSTERS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_ROSTER_STATE;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_RUNNING_HASHES;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_SCHEDULED_COUNTS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_SCHEDULED_ORDERS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_SCHEDULED_USAGES;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_SCHEDULES_BY_EQUALITY;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_SCHEDULES_BY_EXPIRY;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_SCHEDULES_BY_ID;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_SCHEDULE_ID_BY_EQUALITY;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_STAKING_INFO;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_THROTTLE_USAGE;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_TOKENS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_TOKEN_RELATIONS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_TOPICS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_TRANSACTION_RECEIPTS_QUEUE;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_TSS_ENCRYPTION_KEYS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_TSS_MESSAGES;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_TSS_STATUS;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_TSS_VOTES;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_DATA_150;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_DATA_151;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_DATA_152;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_DATA_153;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_DATA_154;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_DATA_155;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_DATA_156;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_DATA_157;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_DATA_158;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_DATA_159;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_FILE;
-import static com.hedera.hapi.block.stream.output.StateIdentifier.STATE_ID_UPGRADE_FILE_HASH;
 import static com.hedera.node.app.records.impl.BlockRecordInfoUtils.HASH_SIZE;
 
 import com.hedera.pbj.runtime.io.buffer.Bytes;
@@ -61,138 +13,12 @@ import java.security.NoSuchAlgorithmException;
  * Utility methods for block implementation.
  */
 public class BlockImplUtils {
-    private static final int UNKNOWN_STATE_ID = -1;
 
     /**
      * Prevent instantiation
      */
     private BlockImplUtils() {
         throw new UnsupportedOperationException("Utility Class");
-    }
-
-    /**
-     * Returns the state id for the given service and state key.
-     *
-     * @param serviceName the service name
-     * @param stateKey the state key
-     * @return the state id
-     */
-    public static int stateIdFor(@NonNull final String serviceName, @NonNull final String stateKey) {
-        final var stateId =
-                switch (serviceName) {
-                    case "AddressBookService" -> switch (stateKey) {
-                        case "NODES" -> STATE_ID_NODES.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "BlockRecordService" -> switch (stateKey) {
-                        case "BLOCKS" -> STATE_ID_BLOCK_INFO.protoOrdinal();
-                        case "RUNNING_HASHES" -> STATE_ID_RUNNING_HASHES.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "BlockStreamService" -> switch (stateKey) {
-                        case "BLOCK_STREAM_INFO" -> STATE_ID_BLOCK_STREAM_INFO.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "CongestionThrottleService" -> switch (stateKey) {
-                        case "CONGESTION_LEVEL_STARTS" -> STATE_ID_CONGESTION_STARTS.protoOrdinal();
-                        case "THROTTLE_USAGE_SNAPSHOTS" -> STATE_ID_THROTTLE_USAGE.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "ConsensusService" -> switch (stateKey) {
-                        case "TOPICS" -> STATE_ID_TOPICS.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "ContractService" -> switch (stateKey) {
-                        case "BYTECODE" -> STATE_ID_CONTRACT_BYTECODE.protoOrdinal();
-                        case "STORAGE" -> STATE_ID_CONTRACT_STORAGE.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "EntityIdService" -> switch (stateKey) {
-                        case "ENTITY_ID" -> STATE_ID_ENTITY_ID.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "FeeService" -> switch (stateKey) {
-                        case "MIDNIGHT_RATES" -> STATE_ID_MIDNIGHT_RATES.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "FileService" -> switch (stateKey) {
-                        case "FILES" -> STATE_ID_FILES.protoOrdinal();
-                        case "UPGRADE_DATA[FileID[shardNum=0, realmNum=0, fileNum=150]]" -> STATE_ID_UPGRADE_DATA_150
-                                .protoOrdinal();
-                        case "UPGRADE_DATA[FileID[shardNum=0, realmNum=0, fileNum=151]]" -> STATE_ID_UPGRADE_DATA_151
-                                .protoOrdinal();
-                        case "UPGRADE_DATA[FileID[shardNum=0, realmNum=0, fileNum=152]]" -> STATE_ID_UPGRADE_DATA_152
-                                .protoOrdinal();
-                        case "UPGRADE_DATA[FileID[shardNum=0, realmNum=0, fileNum=153]]" -> STATE_ID_UPGRADE_DATA_153
-                                .protoOrdinal();
-                        case "UPGRADE_DATA[FileID[shardNum=0, realmNum=0, fileNum=154]]" -> STATE_ID_UPGRADE_DATA_154
-                                .protoOrdinal();
-                        case "UPGRADE_DATA[FileID[shardNum=0, realmNum=0, fileNum=155]]" -> STATE_ID_UPGRADE_DATA_155
-                                .protoOrdinal();
-                        case "UPGRADE_DATA[FileID[shardNum=0, realmNum=0, fileNum=156]]" -> STATE_ID_UPGRADE_DATA_156
-                                .protoOrdinal();
-                        case "UPGRADE_DATA[FileID[shardNum=0, realmNum=0, fileNum=157]]" -> STATE_ID_UPGRADE_DATA_157
-                                .protoOrdinal();
-                        case "UPGRADE_DATA[FileID[shardNum=0, realmNum=0, fileNum=158]]" -> STATE_ID_UPGRADE_DATA_158
-                                .protoOrdinal();
-                        case "UPGRADE_DATA[FileID[shardNum=0, realmNum=0, fileNum=159]]" -> STATE_ID_UPGRADE_DATA_159
-                                .protoOrdinal();
-                        case "UPGRADE_FILE" -> STATE_ID_UPGRADE_FILE.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "FreezeService" -> switch (stateKey) {
-                        case "FREEZE_TIME" -> STATE_ID_FREEZE_TIME.protoOrdinal();
-                        case "UPGRADE_FILE_HASH" -> STATE_ID_UPGRADE_FILE_HASH.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "PlatformStateService" -> switch (stateKey) {
-                        case "PLATFORM_STATE" -> STATE_ID_PLATFORM_STATE.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "RecordCache" -> switch (stateKey) {
-                        case "TransactionReceiptQueue" -> STATE_ID_TRANSACTION_RECEIPTS_QUEUE.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "RosterService" -> switch (stateKey) {
-                        case "ROSTERS" -> STATE_ID_ROSTERS.protoOrdinal();
-                        case "ROSTER_STATE" -> STATE_ID_ROSTER_STATE.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "ScheduleService" -> switch (stateKey) {
-                        case "SCHEDULES_BY_EQUALITY" -> STATE_ID_SCHEDULES_BY_EQUALITY.protoOrdinal();
-                        case "SCHEDULES_BY_EXPIRY_SEC" -> STATE_ID_SCHEDULES_BY_EXPIRY.protoOrdinal();
-                        case "SCHEDULES_BY_ID" -> STATE_ID_SCHEDULES_BY_ID.protoOrdinal();
-                        case "SCHEDULE_ID_BY_EQUALITY" -> STATE_ID_SCHEDULE_ID_BY_EQUALITY.protoOrdinal();
-                        case "SCHEDULED_COUNTS" -> STATE_ID_SCHEDULED_COUNTS.protoOrdinal();
-                        case "SCHEDULED_ORDERS" -> STATE_ID_SCHEDULED_ORDERS.protoOrdinal();
-                        case "SCHEDULED_USAGES" -> STATE_ID_SCHEDULED_USAGES.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "TokenService" -> switch (stateKey) {
-                        case "ACCOUNTS" -> STATE_ID_ACCOUNTS.protoOrdinal();
-                        case "ALIASES" -> STATE_ID_ALIASES.protoOrdinal();
-                        case "NFTS" -> STATE_ID_NFTS.protoOrdinal();
-                        case "PENDING_AIRDROPS" -> STATE_ID_PENDING_AIRDROPS.protoOrdinal();
-                        case "STAKING_INFOS" -> STATE_ID_STAKING_INFO.protoOrdinal();
-                        case "STAKING_NETWORK_REWARDS" -> STATE_ID_NETWORK_REWARDS.protoOrdinal();
-                        case "TOKEN_RELS" -> STATE_ID_TOKEN_RELATIONS.protoOrdinal();
-                        case "TOKENS" -> STATE_ID_TOKENS.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    case "TssBaseService" -> switch (stateKey) {
-                        case "TSS_MESSAGES" -> STATE_ID_TSS_MESSAGES.protoOrdinal();
-                        case "TSS_VOTES" -> STATE_ID_TSS_VOTES.protoOrdinal();
-                        case "TSS_ENCRYPTION_KEYS" -> STATE_ID_TSS_ENCRYPTION_KEYS.protoOrdinal();
-                        case "TSS_STATUS" -> STATE_ID_TSS_STATUS.protoOrdinal();
-                        default -> UNKNOWN_STATE_ID;
-                    };
-                    default -> UNKNOWN_STATE_ID;
-                };
-        if (stateId == UNKNOWN_STATE_ID) {
-            throw new IllegalArgumentException("Unknown state '" + serviceName + "." + stateKey + "'");
-        } else {
-            return stateId;
-        }
     }
 
     /**
