+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/services/MigrationContextImpl.java
@@ -25,6 +25,7 @@ import com.swirlds.state.lifecycle.MigrationContext;
 import com.swirlds.state.lifecycle.StartupNetworks;
 import com.swirlds.state.lifecycle.info.NetworkInfo;
 import com.swirlds.state.merkle.MerkleStateRoot;
+import com.swirlds.state.merkle.NewStateRoot;
 import com.swirlds.state.spi.FilteredWritableStates;
 import com.swirlds.state.spi.ReadableStates;
 import com.swirlds.state.spi.WritableStates;
@@ -72,11 +73,11 @@ public record MigrationContextImpl(
     @Override
     public void copyAndReleaseOnDiskState(@NonNull final String stateKey) {
         requireNonNull(stateKey);
-        if (newStates instanceof MerkleStateRoot.MerkleWritableStates merkleWritableStates) {
+        if (newStates instanceof NewStateRoot.MerkleWritableStates merkleWritableStates) {
             merkleWritableStates.copyAndReleaseVirtualMap(stateKey);
         } else if (newStates instanceof FilteredWritableStates filteredWritableStates
                 && filteredWritableStates.getDelegate()
-                        instanceof MerkleStateRoot.MerkleWritableStates merkleWritableStates) {
+                        instanceof NewStateRoot.MerkleWritableStates merkleWritableStates) {
             merkleWritableStates.copyAndReleaseVirtualMap(stateKey);
         } else {
             throw new UnsupportedOperationException("On-disk state is inaccessible");
