+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/services/OrderedServiceMigrator.java
@@ -33,6 +33,7 @@ import com.swirlds.state.lifecycle.Service;
 import com.swirlds.state.lifecycle.StartupNetworks;
 import com.swirlds.state.lifecycle.info.NetworkInfo;
 import com.swirlds.state.merkle.MerkleStateRoot;
+import com.swirlds.state.merkle.NewStateRoot;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.util.HashMap;
@@ -153,7 +154,7 @@ public class OrderedServiceMigrator implements ServiceMigrator {
                             startupNetworks);
                     // Now commit any changes that were made to the entity ID state (since other service entities could
                     // depend on newly-generated entity IDs)
-                    if (entityIdWritableStates instanceof MerkleStateRoot.MerkleWritableStates mws) {
+                    if (entityIdWritableStates instanceof NewStateRoot.MerkleWritableStates mws) {
                         mws.commit();
                         migrationStateChanges.trackCommit();
                     }
