+++ platform-sdk/swirlds-state-api/src/main/java/com/swirlds/state/spi/WritableQueueStateBase.java
@@ -16,6 +16,7 @@
 
 package com.swirlds.state.spi;
 
+import static com.swirlds.state.StateUtils.computeLabel;
 import static java.util.Objects.requireNonNull;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
@@ -31,16 +32,21 @@ import java.util.function.Predicate;
  * @param <E> The type of element in the queue.
  */
 public abstract class WritableQueueStateBase<E> implements WritableQueueState<E> {
-    /** The state key */
+
+    private final String serviceName;
+
+    /** The state key. */
     private final String stateKey;
+
     /** Each element that has been read. At the moment these are not exposed, but could be. */
     private final List<E> readElements = new ArrayList<>();
-    /** Each element that has been added to the queue, but not yet committed */
+
+    /** Each element that has been added to the queue, but not yet committed. */
     private final List<E> addedElements = new ArrayList<>();
-    /**
-     * Listeners to be notified when the queue changes.
-     */
+
+    /** Listeners to be notified when the queue changes. */
     private final List<QueueChangeListener<E>> listeners = new ArrayList<>();
+
     /**
      * The current index into {@link #addedElements} that we have read from.
      *
@@ -49,13 +55,16 @@ public abstract class WritableQueueStateBase<E> implements WritableQueueState<E>
      * {@link #addedElements} list. This index keeps track of where we have read from it.
      */
     private int currentAddedElementIndex = 0;
+
     /** An iterator from the backing datasource for reading data */
     private Iterator<E> dsIterator = null;
+
     /** The cached most recent peeked element */
     private E peekedElement = null;
 
     /** Create a new instance */
-    protected WritableQueueStateBase(@NonNull final String stateKey) {
+    protected WritableQueueStateBase(@NonNull final String serviceName, @NonNull final String stateKey) {
+        this.serviceName = requireNonNull(serviceName);
         this.stateKey = requireNonNull(stateKey);
     }
 
@@ -120,6 +129,11 @@ public abstract class WritableQueueStateBase<E> implements WritableQueueState<E>
     @Override
     public String getStateKey() {
         return stateKey;
+    }
+
+    // TODO: refactor? (it is duplicated in ReadableQueueStateBase)
+    protected String getLabel() {
+        return computeLabel(serviceName, stateKey);
     }
 
     @Nullable
