+++ platform-sdk/swirlds-state-impl/src/main/java/com/swirlds/state/merkle/disk/OnDiskWritableKVState.java
@@ -30,6 +30,9 @@ import com.swirlds.state.spi.WritableKVStateBase;
 import com.swirlds.state.spi.metrics.StoreMetrics;
 import com.swirlds.virtualmap.VirtualMap;
 import edu.umd.cs.findbugs.annotations.NonNull;
+
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
 import java.util.Iterator;
 import java.util.Objects;
 
@@ -43,7 +46,7 @@ import java.util.Objects;
 public final class OnDiskWritableKVState<K, V> extends WritableKVStateBase<K, V> {
 
     /** The backing merkle data structure */
-    private final VirtualMap virtualMap;
+    private final VirtualMap megaMap;
 
     @NonNull
     private final Codec<K> keyCodec;
@@ -59,28 +62,28 @@ public final class OnDiskWritableKVState<K, V> extends WritableKVStateBase<K, V>
      * @param stateKey     the state key
      * @param keyCodec     the codec for the key
      * @param valueCodec   the codec for the value
-     * @param virtualMap   the backing merkle data structure to use
+     * @param megaMap   the backing merkle data structure to use
      */
     public OnDiskWritableKVState(
-            String stateKey,
+            @NonNull final String serviceName,
+            @NonNull final String stateKey,
             @NonNull final Codec<K> keyCodec,
             @NonNull final Codec<V> valueCodec,
-            @NonNull final VirtualMap virtualMap) {
-        super(stateKey);
+            @NonNull final VirtualMap megaMap) {
+        super(serviceName, stateKey);
         this.keyCodec = keyCodec;
         this.valueCodec = valueCodec;
-        this.virtualMap = requireNonNull(virtualMap);
+        this.megaMap = requireNonNull(megaMap);
     }
 
     /** {@inheritDoc} */
     @Override
     protected V readFromDataSource(@NonNull K key) {
-        final var kb = keyCodec.toBytes(key);
         // FUTURE work: remove legacy hash code
         final int legacyKeyHashCode = Objects.hash(key); // matches OnDiskKey.hashCode()
-        final var value = virtualMap.get(kb, legacyKeyHashCode, valueCodec);
+        final var value = megaMap.get(getMegaMapKey(key), legacyKeyHashCode, valueCodec);
         // Log to transaction state log, what was read
-        logMapGet(getStateKey(), key, value);
+        logMapGet(getLabel(), key, value);
         return value;
     }
 
@@ -89,8 +92,8 @@ public final class OnDiskWritableKVState<K, V> extends WritableKVStateBase<K, V>
     @Override
     protected Iterator<K> iterateFromDataSource() {
         // Log to transaction state log, what was iterated
-        logMapIterate(getStateKey(), virtualMap, keyCodec);
-        return new OnDiskIterator<>(virtualMap, keyCodec);
+        logMapIterate(getLabel(), megaMap, keyCodec);
+        return new OnDiskIterator<>(megaMap, keyCodec);
     }
 
     /** {@inheritDoc} */
@@ -103,28 +106,27 @@ public final class OnDiskWritableKVState<K, V> extends WritableKVStateBase<K, V>
         // If we expect a lot of empty values, Bytes.EMPTY optimization below may be helpful, but
         // for now it just adds a call to measureRecord(), but benefits are unclear
         // final Bytes v = valueCodec.measureRecord(value) == 0 ? Bytes.EMPTY : valueCodec.toBytes(value);
-        virtualMap.put(kb, legacyKeyHashCode, value, valueCodec);
+        megaMap.put(getMegaMapKey(key), legacyKeyHashCode, value, valueCodec);
         // Log to transaction state log, what was put
-        logMapPut(getStateKey(), key, value);
+        logMapPut(getLabel(), key, value);
     }
 
     /** {@inheritDoc} */
     @Override
     protected void removeFromDataSource(@NonNull K key) {
-        final var k = keyCodec.toBytes(key);
         // FUTURE work: remove legacy hash code
         final int legacyKeyHashCode = Objects.hash(key); // matches OnDiskKey.hashCode()
-        final var removed = virtualMap.remove(k, legacyKeyHashCode, valueCodec);
+        final var removed = megaMap.remove(getMegaMapKey(key), legacyKeyHashCode, valueCodec);
         // Log to transaction state log, what was removed
-        logMapRemove(getStateKey(), key, removed);
+        logMapRemove(getLabel(), key, removed);
     }
 
     /** {@inheritDoc} */
     @Override
     public long sizeOfDataSource() {
-        final var size = virtualMap.size();
+        final var size = megaMap.size();
         // Log to transaction state log, size of map
-        logMapGetSize(getStateKey(), size);
+        logMapGetSize(getLabel(), size);
         return size;
     }
 
@@ -140,5 +142,38 @@ public final class OnDiskWritableKVState<K, V> extends WritableKVStateBase<K, V>
         if (storeMetrics != null) {
             storeMetrics.updateCount(sizeOfDataSource());
         }
+    }
+
+    // TODO: test this method
+    // TODO: refactor? (it is duplicated in OnDiskReadableKVState)
+    /**
+     * Generates a key for identifying an entry in the MegaMap data structure.
+     * <p>
+     * The key consists of:
+     * <ul>
+     *   <li>The first 2 bytes: the state ID (unsigned 16-bit, big-endian)</li>
+     *   <li>The remaining bytes: the serialized form of the provided key</li>
+     * </ul>
+     * The state ID must be within [0..65535].
+     * </p>
+     *
+     * @param key the key to serialize and append to the state ID
+     * @return a {@link Bytes} object containing the state ID followed by the serialized key
+     * @throws IllegalArgumentException if the state ID is outside [0..65535]
+     */
+    private Bytes getMegaMapKey(final K key) {
+        final int stateId = getStateId();
+
+        if (stateId < 0 || stateId > 65535) {
+            throw new IllegalArgumentException("State ID " + stateId + " must fit in [0..65535]");
+        }
+
+        final ByteBuffer buffer = ByteBuffer.allocate(2).order(ByteOrder.BIG_ENDIAN);
+        buffer.putShort((short) stateId);
+        final Bytes stateIdBytes = Bytes.wrap(buffer.array());
+
+        final Bytes keyBytes = keyCodec.toBytes(key);
+
+        return stateIdBytes.append(keyBytes);
     }
 }
