+++ platform-sdk/swirlds-state-impl/src/main/java/com/swirlds/state/merkle/queue/QueueNode.java
@@ -28,7 +28,7 @@ import com.swirlds.common.merkle.impl.PartialBinaryMerkleInternal;
 import com.swirlds.common.merkle.utility.DebugIterationEndpoint;
 import com.swirlds.common.utility.Labeled;
 import com.swirlds.fcqueue.FCQueue;
-import com.swirlds.state.merkle.StateUtils;
+import com.swirlds.state.StateUtils;
 import com.swirlds.state.merkle.singleton.StringLeaf;
 import com.swirlds.state.merkle.singleton.ValueLeaf;
 import edu.umd.cs.findbugs.annotations.NonNull;
