--- platform-sdk/swirlds-state-impl/src/main/java/com/swirlds/state/merkle/singleton/ReadableSingletonStateImpl.java
@@ -1,26 +1,0 @@
-/*
- * Copyright (C) 2020-2024 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.swirlds.state.merkle.singleton;
-
-import com.swirlds.state.spi.ReadableSingletonStateBase;
-import edu.umd.cs.findbugs.annotations.NonNull;
-
-public class ReadableSingletonStateImpl<T> extends ReadableSingletonStateBase<T> {
-    public ReadableSingletonStateImpl(@NonNull final String stateKey, @NonNull final SingletonNode<T> node) {
-        super(stateKey, node::getValue);
-    }
-}
